+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.6 - 2020-02-11
+
+### Added
+- ```rails g twirp service``` generates swagger file at public/swagger, options to skip it or set output path 
+- ```rails g twirp service``` generates handler rspec test
+
 ## 0.1.5 - 2020-02-04
 
 ### Added